app/requirements.txt
websockets
pyyaml
argh
watchdog
Pillow
